libstagefright - correct casting of composition time
 - in SampleIterator, findSampleTime needs to correctly cast
    composition time to a signed integer to allow for negative
    values
 - the cast was accidentally removed with 64 bit time stamp change

CRs-Fixed: 360150
(cherry picked from commit 98044856c1513e4c2d7ab9ad7c5362578e66100a)

+++ media/libstagefright/SampleIterator.cpp
@@ -307,7 +307,7 @@ status_t SampleIterator::findSampleTime(
 
     *time = mTTSSampleTime + (uint64_t)mTTSDuration * ((uint64_t)sampleIndex - (uint64_t)mTTSSampleIndex);
 
-    *time = *time + mTable->getCompositionTimeOffset(sampleIndex);
+    *time = *time + (int32_t)mTable->getCompositionTimeOffset(sampleIndex);
 
     return OK;
 }
